moved all tests into specific test folder, for better expandabiliy
