+++ canvas_project/editor/tests/test_views.py
